Model: all-MiniLM-L6-v2
What is the primary responsibility of employees regarding the Code of Ethics?
Compliance with the Code of Ethics.

Reference: https://policy.uconn.edu/2011/06/02/use-of-students-in-outside-employment/
What may happen if an employee violates the Code of Ethics?
The Office of State Ethics has jurisdiction to interpret and enforce the Code. Violations may result in a formal complaint proceeding filed against the employee and sanctions of up to $10,000 per violation. The entire Code and regulations, as well as a summary of these rules, may be found at the

Reference: https://policy.uconn.edu/2011/05/24/guide-to-the-state-code-of-ethics/
What should an employee do if they have questions about the provisions of the Code?
The Office of State Ethics has jurisdiction to interpret and enforce the Code. Violations may result in a formal complaint proceeding filed against the employee and sanctions of up to $10,000 per violation. The entire Code and regulations, as well as a summary of these rules, may be found at the

Reference: https://policy.uconn.edu/2011/05/24/guide-to-the-state-code-of-ethics/
Are employees allowed to accept gifts from prohibited donors?
In general, employees are prohibited from accepting gifts, discounts or gratuities of any kind from prohibited donors: (1) doing business with or seeking to do business with the University; (2) directly regulated by the University; (3) known to be a registered lobbyist or a lobbyist’s representative, or; (4) pre-qualified under Conn. Gen. Stat. §4a-
100. A list of registered lobbyists can be found on the web site of the Office of State Ethics. Certain items are excluded from the definition of “gift,” including: items offered to the public at large (for example, trinkets provided at an open house), items valued at under $10, food and beverage up to $50 in a calendar year from each donor and training for a product purchased by the University provided such training is offered to all customers of that vendor. “Gifts to the state” are also permitted as long as the gifts facilitate University actions or functions. If an employee is offered a benefit from someone other than the prohibited donors listed above, and the benefit is offered because of the employee’s position at the University, the total value of benefits received from one source in a year must not exceed $
100. Additionally, supervisors may only accept gifts valued at no more than $100 from a subordinate; a subordinate may only accept gifts valued at no more than $100 from his/her supervisor. This provision not only applies to direct supervisors and subordinates, but to any individual up or down the chain of command. Questions regarding specific facts and circumstances surrounding various gift-giving scenarios should be directed to the University’s Ethics Liaison or the Office of State Ethics.

Reference: https://policy.uconn.edu/2011/05/24/guide-to-the-state-code-of-ethics/